cfpb-backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
python-dotenv
scikit-learn
scipy
joblib
pandas
numpy
faiss-cpu
sentence-transformers
spacy
vaderSentiment
nltk
transformers
torch
gradio
rouge-score
wordcloud
umap-learn